refactor fractal flame logic; use for webcam; wip webcam
