+++ src/custom-typings/three.d.ts
@@ -1,0 +1,7 @@
+import * as THREE from "three";
+
+// changes in r88
+
+declare module 'THREE' {
+    export const AxesHelper = THREE.AxisHelper;
+}
